refactor(cmd): merge audit_tools + audit_output into audit_surface_quality
Combines the metadata-quality audit (naming, annotations, schema shape —
formerly audit_tools) and the output-quality audit (OutputSchema, Returns/See-
also, Title — formerly audit_output) behind a single -view flag.

  cmd/audit_surface_quality/
  ├── main.go                 # -view={metadata,output,all}, shared listTools
  ├── metadata_audit.go       # 8 metadata checks + register-meta inventory
  ├── metadata_register_meta.go
  ├── output_audit.go         # 5 output-quality checks
  └── *_test.go               # all tests migrated in-package

Equivalence: each -view produces structurally-equivalent markdown to the former
standalone binary (verified against plan/baselines/phase0/, modulo the volatile
Generated: timestamp). The shared listTools applies LockdownInputSchemas (the
metadata binary's stricter behavior); output checks don't depend on lockdown, so
no behavioral change.

Makefile keeps audit-tools/audit-output as -view= wrappers for backward compat.
net cmd/ count: 22 → 21.
