+++ docs/testing/testing.md
@@ -18,10 +18,10 @@
 
 | Metric                                                |  Value |
 | ----------------------------------------------------- | -----: |
-| Total test functions                                  | 11,562 |
-| Unit test functions                                   | 11,280 |
+| Total test functions                                  | 11,560 |
+| Unit test functions                                   | 11,278 |
 | E2E test functions                                    |    282 |
-| cmd test functions                                    |    870 |
+| cmd test functions                                    |    868 |
 | Test files (internal/)                                |    453 |
 | Test files (cmd/)                                     |     61 |
 | Test files (test/e2e/suite/)                          |    139 |
@@ -35,7 +35,7 @@
 
 | Pattern                                |  Count |     % |
 | -------------------------------------- | -----: | ----: |
-| `TestFunc_Scenario` (2-part)           | 10,301 | 89.1% |
+| `TestFunc_Scenario` (2-part)           | 10,299 | 89.1% |
 | `TestFunc` (no underscore)             |    933 |  8.1% |
 | `TestFunc_Scenario_Expected` (3+ part) |    328 |  2.8% |
 
@@ -49,8 +49,8 @@
 | Tools orchestration     |            283 |         13 | registration, meta-tool dispatch, safe mode, validation, markdown, and routing tests            |
 | Tool sub-packages (174) |          8,263 |        348 | domain-specific GitLab tool handlers                                                            |
 | E2E integration         |            282 |        139 | build-tagged real GitLab integration suite                                                      |
-| cmd packages            |            870 |         61 | server entry point and developer command utilities                                              |
-| **Total**               |     **11,562** |    **653** |                                                                                                 |
+| cmd packages            |            868 |         61 | server entry point and developer command utilities                                              |
+| **Total**               |     **11,560** |    **653** |                                                                                                 |
 
 ### Core Packages
